requirements.txt
google-generativeai>=0.3.0
openai>=1.0.0
python-dotenv>=1.0.0
python-docx>=1.1.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
jupyter>=1.0.0
ipykernel>=6.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
